Added state machine and bubble sort

+++ Main/LedStrip.hpp
@@ -3,20 +3,20 @@
 
 class LedStrip {
   public:
-    LedStrip(uint16_t pixelCount, uint8_t pin);
+    LedStrip(int32_t pixelCount, uint8_t pin);
     LedStrip(const LedStrip &) = delete;
-    uint16_t PixelCount();
-    RgbColor GetColor(uint16_t pixel);
-    void SetColor(uint16_t pixel, RgbColor color);
+    int32_t PixelCount();
+    RgbColor GetColor(int32_t pixel);
+    void SetColor(int32_t pixel, RgbColor color);
     void SetColors(RgbColor color);
-    void SetColors(uint16_t first_pixel, uint16_t last_pixel, RgbColor color);
+    void SetColors(int32_t first_pixel, int32_t last_pixel, RgbColor color);
     void Gradient(RgbColor first_color, RgbColor last_color);
-    void Gradient(uint16_t first_pixel, uint16_t last_pixel,
-                  RgbColor first_color, RgbColor last_color);
+    void Gradient(int32_t first_pixel, int32_t last_pixel, RgbColor first_color,
+                  RgbColor last_color);
     void Mirror();
-    void Mirror(uint16_t first_pixel, uint16_t last_pixel);
-    void Rotate(int16_t offset);
-    void Rotate(uint16_t first_pixel, uint16_t last_pixel, int16_t offset);
+    void Mirror(int32_t first_pixel, int32_t last_pixel);
+    void Rotate(int32_t offset);
+    void Rotate(int32_t first_pixel, int32_t last_pixel, int32_t offset);
     void Update();
 
   private:
